+++ spec/jobs/import/sync_repositories_job_spec.rb
@@ -23,7 +23,8 @@ describe Import::SyncRepositoriesJob, type: :service do
     before do
       company.configuration.assign_attributes(
         work_start_time: DateTime.new(2023, 1, 1, 9, 0),
-        work_end_time: DateTime.new(2023, 1, 1, 18, 0)
+        work_end_time: DateTime.new(2023, 1, 1, 18, 0),
+        work_time_zone: 'Moscow'
       )
       company.save!
     end
@@ -31,6 +32,18 @@ describe Import::SyncRepositoriesJob, type: :service do
     context 'with current time inside working time' do
       before do
         allow(DateTime).to receive(:now).and_return(DateTime.new(2023, 1, 2, 10, 0))
+      end
+
+      it 'calls service' do
+        job_call
+
+        expect(import_object).to have_received(:call).with(company: company)
+      end
+    end
+
+    context 'with current time inside working time with timezone offset' do
+      before do
+        allow(DateTime).to receive(:now).and_return(DateTime.new(2023, 1, 2, 8, 0))
       end
 
       it 'calls service' do
@@ -54,7 +67,7 @@ describe Import::SyncRepositoriesJob, type: :service do
 
     context 'with current time outside working time' do
       before do
-        allow(DateTime).to receive(:now).and_return(DateTime.new(2023, 1, 2, 7, 0))
+        allow(DateTime).to receive(:now).and_return(DateTime.new(2023, 1, 2, 5, 0))
       end
 
       it 'does not call service' do